Move helpers in CLI module to the internal package
